+++ Programm/ProgressBar.cpp
@@ -6,16 +6,16 @@
 #include <algorithm>
 #include <iostream>
 
-ProgressBar::ProgressBar(std::string prefix, size_t max) : max {max} {
+ProgressBar::ProgressBar(std::string prefix, unsigned max) : max {max} {
     step_size = std::max(1ul, max / max_steps);
     std::cout << prefix << " …" << std::endl;
 
     std::cout << "[";
-    for (int i = 0; i < max_steps; i++) {
+    for (unsigned i = 0; i < max_steps; i++) {
         std::cout << "-";
     }
     std::cout << "]" << std::flush;
-    for (int i = 0; i < max_steps + 1; i++) {
+    for (unsigned i = 0; i < max_steps + 1; i++) {
         std::cout << "\b";
     }
     std::cout << std::flush;
@@ -25,8 +25,8 @@ ProgressBar::~ProgressBar() {
     close();
 }
 
-void ProgressBar::print_update(int value) {
-    int target_steps = value * max_steps / max;
+void ProgressBar::print_update(unsigned value) {
+    unsigned target_steps = value * max_steps / max;
 
     for (; target_steps > printed_steps; printed_steps++) {
         std::cout << "=";
